requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.1
djangorestframework==3.12.2
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
